+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.3.1
+
+- Fix bug: remove test code in bin.
+- Change default `output` to `./download/video`.
+- params `ext` have abbreviation `e`.
+
 ## 1.3.0
 
 - Support windows.